# openapi.yaml
openapi: 3.1.0
info:
  title: Parma Growth Operator
  version: 1.1.0
servers:
  - url: https://supportive-stillness-production-ec37.up.railway.app
paths:
...
  /tools/google/campaign/{id}/intelligence:
    get:
      operationId: getGoogleCampaignIntelligence
      summary: Get complete read-only Google Ads campaign diagnostics
      description: Returns configured and observed read-only Google Ads campaign diagnostics including schedule, targeting criteria, budget sharing, bidding strategy, status reasons, policy signals, negative keywords and RSA analysis. days=0 reads today; days=1 reads yesterday; read_mode=today_intraday explicitly reads today in account timezone.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 90
            default: 30
        - name: read_mode
          in: query
          required: false
          schema:
            type: string
            enum: [historical, today_intraday]
            default: historical
      responses:
        "200":
          description: Complete read-only campaign intelligence
        "400":
          description: Invalid campaign id or date range
        "401":
          description: Unauthorized
        "500":
          description: Google Ads query failed safely
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key